Enable use of focus in rspec tests
https://www.relishapp.com/rspec/rspec-core/v/3-4/docs/filtering/inclusion-filters

When working on a new feature or when making changes to an area of the code, its handy to just run the tests that are relevant. This speeds up development, whilst still allowing you to run the ful suite later.

Rspec comes with a feature called 'Inclusion filters' which provides a handy way to do this. This change simply updates the Rspec config to allow us to use the `filter: true` option in specs.