Copy example secrets to secrets.yml in .travis.yml
Updated the `.travis.yml` config file to ensure `secret_key_base` is set.

After adding a test for to check the response headers, Travis would no longer pass the build because it was expecting the secret key to have been set. Though we set an ENV VAR in the `.travis.yml` file called `SECRET_KEY_BASE` it does not appear rails was seeing it, as it was expecting to pull it from a `config/secrets.yml` file which this project doesn't actually have checked in.

Hence this change adds a line to the Travis config file to simple copy `config/secrets.example.yml` to `config/secrets.yml`.