Update to latest Content-Security-Policy impl.
https://eaflood.atlassian.net/browse/PM-286

This is linked to an update to the Content Security Policy header, to cover content sourced from Google which are required for the analytics (see PR https://github.com/DEFRA/pafs_core/pull/131). That PR not only updates the CSP, but switches away from us setting the CSP to using the [Secure Headers](https://github.com/twitter/secureheaders) gem from **Twitter** to do it.

Going forward this will give us the ability to set other security related headers, plus it includes the ability to work with nonce values for inline scripts.

So the changes here are what are required to work with an updated **pafs_core**.