Update areas.csv from 16 to 14 plus missing RMAs
https://eaflood.atlassian.net/browse/PM-151

The EA areas have changed from 16 to 14 so `areas.csv` which seeds the database upon first install needs to be updated to reflect this.

It also now includes those RMA's which exist, but not currently have a project in the system. Hence though the number of areas as reduced, we have more lines in the file.