Initialize feedback links in config to real values
https://eaflood.atlassian.net/browse/PM-310

We have a feedback link in the header and one when the proposal is finally submitted. Prior to this change these links didn't actually point to anything real.

The links come from config values set in an initializer.

This change updates the `pafs_core_config.rb` initializer to ensure the correct, real values are used.