Remove turbolinks from project
We recently found that the radio buttons were not correctly highlighting when selected (post 0e43cee). This appeared to be a recent change following running a bundle update against the project.

Initially we looked at the GOVUK gems which are responsible for the majority of the lyaout and styling, and spotted they had indeed been updated. However reverting them to previous versions did not seem to resolve the issue.

It was then we instead looked at whether Turbolinks was at fault (the fact that a forced reload of the page was the initial clue). By removing Turbolinks from the project we found the problem went away. We also spotted that Turbolinks had gone from version 2 to version 5 during the recent bundle update so we are now confident it is the root cause.

In order to keep the latest Turbolinks we would need to amend the document load call within the template, which is diffcult as it comes from the GOV.UK gem. We could instead have gone back to the previous version and locked the project to it. However after some discussion we felt it would be better to go with removing turbolinks as this is not a project that will greatly benefit from using it, and we run the risk of wasting time trying to solve similar issues in the future.